+++ ocean_lib/assets/credentials.py
@@ -5,7 +5,7 @@
 from typing import Optional
 
 from enforce_typing import enforce_types
-from ocean_lib.common.agreements.consumable import ConsumableCodes, MalformedCredential
+from ocean_lib.agreements.consumable import ConsumableCodes, MalformedCredential
 
 
 class AddressCredential:
