+++ ocean_lib/assets/test/test_utils.py
@@ -5,6 +5,7 @@
 from unittest.mock import patch
 
 import pytest
+from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.assets.did import DID
 from ocean_lib.assets.trusted_algorithms import (
     add_publisher_trusted_algorithm,
@@ -14,7 +15,6 @@ from ocean_lib.assets.trusted_algorithms import (
     remove_publisher_trusted_algorithm,
     remove_publisher_trusted_algorithm_publisher,
 )
-from ocean_lib.common.agreements.service_types import ServiceTypes
 from tests.resources.ddo_helpers import (
     get_sample_algorithm_ddo,
     get_sample_ddo,
@@ -127,18 +127,14 @@ def test_utilitary_functions_for_trusted_algorithm_publishers(publisher_ocean_in
     assert len(new_publisher_trusted_algo_publishers) == 2
 
     # remove an existing algorithm to publisher_trusted_algorithms list
-    new_publisher_trusted_algo_publishers = (
-        remove_publisher_trusted_algorithm_publisher(
-            ddo, "0xABC", publisher_ocean_instance.config.metadata_cache_uri
-        )
+    new_publisher_trusted_algo_publishers = remove_publisher_trusted_algorithm_publisher(
+        ddo, "0xABC", publisher_ocean_instance.config.metadata_cache_uri
     )
 
     assert len(new_publisher_trusted_algo_publishers) == 1
 
     # remove a trusted algorithm that does not belong to publisher_trusted_algorithms list
-    new_publisher_trusted_algo_publishers = (
-        remove_publisher_trusted_algorithm_publisher(
-            ddo, "0xaaaa", publisher_ocean_instance.config.metadata_cache_uri
-        )
+    new_publisher_trusted_algo_publishers = remove_publisher_trusted_algorithm_publisher(
+        ddo, "0xaaaa", publisher_ocean_instance.config.metadata_cache_uri
     )
     assert len(new_publisher_trusted_algo_publishers) == 1
