+++ ocean_lib/assets/trusted_algorithms.py
@@ -6,9 +6,9 @@ import json
 from typing import Optional, Union
 
 from enforce_typing import enforce_types
+from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.assets.asset import V3Asset
 from ocean_lib.assets.asset_resolver import resolve_asset
-from ocean_lib.common.agreements.service_types import ServiceTypes
 from ocean_lib.utils.utilities import create_checksum
