+++ ocean_lib/ocean/ocean_assets.py
@@ -14,12 +14,12 @@ from typing import Optional, Tuple, Type, Union
 from enforce_typing import enforce_types
 from eth_account.messages import encode_defunct
 from eth_utils import add_0x_prefix, remove_0x_prefix
+from ocean_lib.agreements.consumable import AssetNotConsumable, ConsumableCodes
+from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.assets.asset import V3Asset
 from ocean_lib.assets.asset_downloader import download_asset_files
 from ocean_lib.assets.asset_resolver import resolve_asset
 from ocean_lib.assets.did import did_to_id
-from ocean_lib.common.agreements.consumable import AssetNotConsumable, ConsumableCodes
-from ocean_lib.common.agreements.service_types import ServiceTypes
 from ocean_lib.common.aquarius.aquarius import Aquarius
 from ocean_lib.common.aquarius.aquarius_provider import AquariusProvider
 from ocean_lib.config import Config
