+++ ocean_lib/services/service.py
@@ -12,7 +12,7 @@ from typing import Any, Dict, Optional
 from urllib.parse import urlparse
 
 from enforce_typing import enforce_types
-from ocean_lib.common.agreements.service_types import ServiceTypes, ServiceTypesIndices
+from ocean_lib.agreements.service_types import ServiceTypes, ServiceTypesIndices
 from ocean_lib.data_provider.data_service_provider import DataServiceProvider
 
 logger = logging.getLogger(__name__)
